ADD: dry-run option to both rename.py and replace.py

+++ rename.py
@@ -15,11 +15,13 @@ def main():
     parser.add_argument("path", help="the path to start renaming files recursively")
     parser.add_argument("old_pattern", help="the old pattern to be replaced by")
     parser.add_argument("new_pattern", help="the new pattern to be replaced with")
+    parser.add_argument('--dry_run', action='store_true', help="show what would happen without doing it")
+
     args = parser.parse_args()
 
     # rename_files(args.path, new_name=new_name7)
     f = multi_replace_curry([(args.old_pattern, args.new_pattern)])
-    rename_files(args.path, new_name_func=f)
+    rename_files(args.path, new_name_func=f, dry_run=args.dry_run)
 
 
 def new_name1(name):
@@ -55,7 +57,7 @@ def new_name7(name):
     return re.sub(r"Track \d+ -", "Track", name)
 
 
-def rename_files(directory, new_name_func):
+def rename_files(directory, new_name_func, dry_run):
     for file_name in os.listdir(directory):
         old_file_path = os.path.join(directory, file_name)
         if os.path.isdir(old_file_path):
@@ -66,7 +68,9 @@ def rename_files(directory, new_name_func):
             new_file_path = os.path.join(directory, new_file_name)
 
             logging.info("Renaming '%s' to '%s'" % (old_file_path, new_file_path))
-            os.rename(old_file_path, new_file_path)
+
+            if not dry_run:
+                os.rename(old_file_path, new_file_path)
         else:
             logging.debug("Ignoring '%s'" % old_file_path)
